Update email config for development env (#198)
The current configuration relies on [letter_opener](https://github.com/ryanb/letter_opener) to see emails when working in `development` mode.

That's fine if you are running the app directly on the host. But we also want to support the option to run the app in tools like [vagrant](https://www.vagrantup.com/) and [docker](https://www.docker.com/) where this no longer works.

So this updates the config to assume you have [Mailcatcher](https://mailcatcher.me/) running when working locally. This better represents production and gives us more flexibility on where we can run the app.

N.B. This config and approach are taken from Waste Carriers, Waste Exemptions and Flood Risk Activity Exemptions which are also Rails-based services.