Integrate the project with CI (#229)
The main focus is on getting the project to build and run tests successfully in Travis. Then we can also run SonarCloud analysis as well as ensuring we record our test coverage.

A key issue with this project is its reliance on Selenium-based integration tests. So the Travis build config is a little more complex than normal in order to support this.

And finally, badges of course!