Add snyk badge to README
We have also hooked the project up to [Snyk](https://app.snyk.io/org/cruikshanks/projects) because we have a dependence on a number of npm packages as well as ruby gems.

Hakiri only supports Ruby apps so this will be a helpful addition in monitoring the security of the app.