Integrate RSpec for unit testing (#267)
Currently the project uses [minitest](https://github.com/seattlerb/minitest) as its test framework. Nothing against Minitest; its the default for Rails for example. It's just that _every other_ Defra ruby based project uses RSpec, and the [new teams](https:/github.com/DEFRA/sroc-service-team) experience is only with RSpec.

So not only to allow us to maintain it but to make it easier for Defra colleagues to help out if needed we feel we need to bring in RSpec.

This change will integrate RSpec and get it up and running whilst checking we don't break any existing stuff.

We'll then convert one of the existing tests across as an example in a subsequent PR. The intention is all new tests will be done in RSpec and existing tests will be converted

- at any point, we need to make changes to them
- as and when time permits

Also covered in this change are

* Update .gitignore

Needed to add `/spec/examples.txt` to it but then spotted duplicates. So took the time to update it to be consistent with our other rails based projects.

* Fix Simplecov and add example RSpec test

To confirm things were working in RSpec we found a basic helper method we could add a test for. This didn't appear to be covered by Minitest so it was an added bonus.

Whilst doing this and ensuring when we run both RSpec and Minitest we get a combined test coverage it highlighted the previous Simplecov setup was incorrect. Looking at the previous coverage report

- some files were missing which should have been covered
- files in the config folder were included in the metrics (don't normally cover this folder)
- all the `test/*_test.rb` files were being included. As they are run under 'test' it was logging 100% test coverage for each file

The end result was that even though more files were being tested, they were the wrong files to be tested and they were reporting back incorrect coverage data. This is why the Simplecov reporting has changed from

`91.58% - 217 files in total. 7260 relevant lines. 6649 lines covered and 611 lines missed`

to

`55.93% - 170 files in total. 5183 relevant lines. 2899 lines covered and 2284 lines missed`