Example API Connector implementation (#268)
Example API Connector implementation

https://github.com/DEFRA/sroc-tcm-admin/pull/192

This is an example of how the work in PR #192 could be reworked to use modules. Our blocker was testing, and the fact we only really knew how to go about it using RSpec. So we needed [RSpec integrated first](https://github.com/DEFRA/sroc-tcm-admin/pull/267) before we could tackle this example.

After completing the example, we agreed this was the version we wished to continue with. Hence we have now merged it in place of PR #192