Update seed script to create test users (#305)
We currently have an issue with the seeding of user accounts. We rely on a real email account to be used because we generate an unknowable password as part of the seeding. We need a real account so we can then the password reset mechanism to actually gain access.

This means if you were to seed a fresh environment, you could never actually use the `system@example.com` user unless you were able to intercept the email (only possible locally). The current seed design also makes it hard to write automated acceptance tests as we ideally need known users to authenticate with.

So the primary purpose for this change is to start seeding known user accounts based on roles and regimes. All will use the same known password taken from an environment variable.

* Update seed script to create users

After some checking, I found the service has a dependency on `system@example.com`. It seems to be a user reserved for actions the system takes automatically so must be created.

Other than that this updates `db/seeds.rb` with logic that will create users based on a `users.json` file. We only need to add new users to that and `seeds.rb` will pick them up.

The current set of users ensures

- we create a billing user-specific to each regime
- we create a user for each role and add them to all regimes to allow testing across them

Wil leave this PR here. Our intent once approved and merged is to follow it up with a change that ensures seeding is only allowed when

- the environment is not production
- if production, a special env var that permits seeding has been set