Update project to use ruby 2.7.1 (#340)
https://trello.com/c/gKnSb5XO

The current ruby version is very out of date and no longer supported or maintained. This includes not receiving any security fixes.

Currently, it's 2.4.1 (March 2017). The latest is 2.7.2 (October 2020). However, 2.7.2 still isn't listed in a number of build tools plus we have a number of other services using 2.7.1 right now. So we are sticking with 2.7.1 to be consistent.

This covers everything needed to get the service up and running with ruby 2.7.1.

Note - Re-running the migrations locally caused the schema.rb to be re-generated. This dropped the frozen string literal magic comment. However, that's not an issue as we actually have ignoring schema.rb setup in our default defra-ruby-style project.