Upgrade project to use Rails 6 (#341)
https://trello.com/c/PZ2Q3LIa

The current rails version is very out of date and no longer supported or maintained. Currently, it's 5.1.1 (May 2017). The latest v5 version is 5.2.4.3 (May 2020). Rails v5.2 is still receiving security fixes at least. But the project would still need to be updated to it.

The latest version of Rails is 6.0.3.4 (October 2020). Only Rails v6.0 is getting bug fixes as well as security updates.

This change covers updating the project to use Rails 6 along with any other changes needed to support it.

It meant running a `bundle update` to rebuild the `Gemfile.lock`. After this the app still built and tests passed. But we were flooded with deprecation warnings that needed to be resolved next.

## Resolve update_attributes warning

```
DEPRECATION WARNING: update_attributes is deprecated and will be removed from Rails 6.1 (please, use update instead)
```

## Fix dangerous query method warning

```
DEPRECATION WARNING: Dangerous query method (method whose arguments are used as raw SQL) called with non-attribute argument(s): "string_to_array(code, '.')::int[] asc". Non-attribute arguments will be disallowed in Rails 6.1. This method should not be called with user-provided values, such as request parameters or model attributes. Known-safe values can be passed by wrapping them in Arel.sql().
```

## Remove coffeescript gem

```
DEPRECATION WARNING: Single arity template handlers are deprecated. Template handlers must
now accept two parameters, the view object and the source for the view object.
Change:
  >> Coffee::Rails::TemplateHandler.call(template)
To:
  >> Coffee::Rails::TemplateHandler.call(template, source)
```

## Fix 'last argument as keyword parameters' warning

```
warning: Using the last argument as keyword parameters is deprecated; maybe ** should be added to the call
```

See https://piechowski.io/post/last-arg-keyword-deprecated-ruby-2-7/ for reference

## Fix 'Passing the keyword argument' warning

```
warning: Passing the keyword argument as the last hash parameter is deprecated
```

See https://bloggie.io/@kinopyo/how-to-fix-ruby-2-7-warning-using-the-last-argument-as-keyword-parameters-is-deprecated for reference

## Fix 'has_style? is deprecated' warning

```
DEPRECATED: has_style? is deprecated, please use matches_style?
```

## Fix ':driver_opts is deprecated'

```
WARN Selenium [DEPRECATION] :driver_opts is deprecated. Use :service with an instance of Selenium::WebDriver::Service instead.
```

## Fix code coverage

With everything being updated it included bumping simplecov to its latest (actually v0.19). The problem is from v0.18 they changed the format of the output file which broke a number of CI services including SonarCloud.

So lock it to ensure we are still using a format SonarCloud recognises.

## Clear up confusion about app servers

Both puma and passenger are app servers for rails. But we are referencing both in our Gemfile.

Currently, we are using passenger in production as that is what web-ops has configured the environment for. Locally Rails 5 switched from webrick to puma (it actually switched to it for everything!)

This helps clear up some of the confusion by referencing the right app server in the right group to make it clear which is used where.

## Update puma config

Based on the fact we are not running puma in production this sets the minimum thread count to be 1 as that is all we actually want to run in development/test to aid monitoring and debugging.

## Update puma to latest

Puma is actually on version 5 whilst we were pulling in v3. This updates it to the latest.

Note - we do now get this message when running unit tests `/home/vagrant/.rbenv/versions/2.7.1/lib/ruby/gems/2.7.0/gems/puma-5.0.4/lib/puma/server.rb:94: warning: deprecated Object#=~ is called on Proc; it always returns nil` but expect that to be fixed in a release soon.

## Remove tzinfo-data gem

This gem supports running the app on Windows. As this is an internal app specific to the EA we don't expect anyone else to be making use of it. Plus the whole development team and our environments are all running some nix version and not windows.

Removing this gem gets rid of a bunch of warning messages from different places in the app lifecycle.