Remove redundant check rails rake task (#382)
Whilst migrating some existing Jenkins jobs to Jenkins pipelines we came across one intended to check that the app is up and running (it doesn't work!)

Whilst doing this we also came across `tcm:check_rails_running` within the app itself. This will only work locally because when the app is deployed in an AWS environment using Capistrano the PID files go into a shared directory.

We should add, we can also just make a request ourselves to confirm.

So this change removes the redundant, non-working rake task.