Talk to the rules service directly (#406)
Currently, this service talks to the Rules Service via [sroc-charging-service](https://github.com/DEFRA/sroc-charging-service).

The **sroc-charging-service** is built in Node so an entirely different technology to the TCM. It also talks about being an abstraction and simplification wrapper for talking to the Rules Service. In fact, we know it was used to solve the problem of getting the TCM to see the Rules Service when deployed to our AWS environments.

That last issue _is_ solvable. The current team have experience working on a number of other services deployed to the same AWS environments which all can communicate with external services.

If we can get the TCM talking to the rules service directly, we can drop the overhead of maintaining the **sroc-charging-service** project and all the overhead that goes with it (updates and patches, deployment, Jenkins jobs, documentation etc).

This change covers everything needed to make that happen.