Fix permissions on import fixture (and force add)
Spotted that the permissions on our import fixture file were different to others, specifically only the current user had read permissions.

```
-rw------- 1 vagrant vagrant 1019 Apr 19 15:42 cfdti999.dat.csv
```

Compare that to the existing fixtures

```
-rw-r--r-- 1 vagrant vagrant  112 Feb  5 14:55 create_bill_run_success.json
-rw-r--r-- 1 vagrant vagrant 1664 Feb  5 14:55 list_bill_runs_success.json
```

Something wasn't right. So, we amended the permissions on the file though this didn't seem enough to make git pick up the change. Hence a tweak to the content to force it.