Add test for a file with missing data
This was an attempt to provide coverage for when `transaction_file&.valid?` returned `false` in `FileImportService`. We soon realised though this would never happen.

This is because the previous `importer.import(in_file.path, File.basename(f))` call which generates the `transaction_file` (actually an instance of `TransactionHeader`) uses `TransactionHeader.create!()` and `header.save!`. These both will result in an error being thrown if anything is invalid with the data.

So, we'll never get a `TransactionHeader` returned that has invalid data, which means `valid?()` will never return `false`.