Migrate File Check Job to rake (#412)
> Part of removing redis and resque dependencies
>
> The service currently uses [Resque](https://github.com/resque/resque) for managing background jobs including email. We use [whenever](https://github.com/javan/whenever) to schedule adding 'jobs' to the queue, which means adding a record in [redis](https://redis.io/). Registered resque 'workers' then pick up and process the actual jobs.
>
> This would be great if we were dealing with thousands of background tasks. But we're not. We are dealing with just 3. Recently we have been caught out a number of times with issues owing to either jobs not logging, workers not being registered, and confusion about the process overall.
>
> There is just no need for all this additional complexity when we can replicate what other services have done and which this is already doing; just have **whenever** via cron run the task directly.

This change focuses on the `FileCheckJob` and migrating it to be run from a [rake](https://github.com/ruby/rake) task.

** Notes

The key change was migrating the job to a service and therefore sticking with the current convention of inheriting from  `ServiceObject`. This meant some minor changes plus the need to ensure we support `success?` and `failed?` properly.

None of the existing logging actually got recorded anywhere because it was using `Rails.logger`. Plus it was not suitable for something we will now run from a rake task and expect output to STDOUT. This explains the switch to just using `puts`. We also tweaked the messages to include the file being imported as we felt this information was critical.

With this change, we also took the opportunity to add some tests (there was nothing covering it before). To support this we needed to add FactoryBot and DatabaseCleaner. Both are tools we use in all our other projects which bring this one more inline. Specifically, we needed DatabaseCleaner because we found that the existing Minitest suite was leaving records behind in the DB which then conflicted with our tests in RSpec. Should the day ever come when everything is in RSpec it may no longer be needed.

A final thing to note about the tests. The previous decision to use a local file store when the app runs in `DEVELOPMENT` or `TEST` made stubbing some parts of the import process difficult. We've opted to stick with a process of copying fixtures to where they need to be for `FileImportService` to work under test conditions. This meant we needed to create a number of fixture files. Ideally, we'd use AWS S3 irrespective of what environment we are working in, which would also make stubbing a response possible instead of copying files. But that kind of change was beyond the scope of this PR.